chore: pull plugin into own file, remove explicit return type
