+++ src/index.ts
@@ -1,4 +1,5 @@
 import "./globals";
 
 export type { GetLoadContextFunction, RequestHandler } from "./server";
-export { createRequestHandler, remixFastifyPlugin } from "./server";
+export { createRequestHandler } from "./server";
+export { remixFastifyPlugin } from "./plugin";
