+++ src/server.ts
@@ -1,8 +1,4 @@
-import path from "path";
-
-import fp from "fastify-plugin";
-import fastifyStatic from "@fastify/static";
-import type { FastifyPluginAsync, FastifyReply, FastifyRequest } from "fastify";
+import type { FastifyRequest, FastifyReply } from "fastify";
 import type {
   AppLoadContext,
   ServerBuild,
@@ -16,70 +12,6 @@ import {
   Headers as NodeHeaders,
   Request as NodeRequest,
 } from "@remix-run/node";
-
-interface PluginOptions {
-  mode?: string;
-  build?: ServerBuild;
-  // Same as assetsBuildDirectory in remix.config.js, but absolute or relative to this file
-  assetsBuildDirectory?: string;
-  // Must match the same setting in remix.config.js
-  publicPath?: string;
-}
-
-const remixFastify: FastifyPluginAsync<PluginOptions> = async (
-  fastify,
-  options = {}
-) => {
-  let fullOptions = Object.assign(
-    {
-      mode: process.env.NODE_ENV,
-      assetsBuildDirectory: path.resolve(process.cwd(), "public", "build"),
-      publicPath: "/build/",
-    },
-    options
-  );
-
-  if (!fullOptions.build) {
-    throw new Error("Must provide a build");
-  }
-
-  if (!fastify.hasContentTypeParser("*")) {
-    fastify.addContentTypeParser("*", (_request, payload, done) => {
-      let data = "";
-      payload.on("data", (chunk) => {
-        data += chunk;
-      });
-      payload.on("end", () => {
-        done(null, data);
-      });
-    });
-  }
-
-  await fastify.register(fastifyStatic, {
-    root: fullOptions.assetsBuildDirectory,
-    prefix: "/",
-    wildcard: false,
-    maxAge: 31536000,
-    dotfiles: "allow",
-  });
-
-  await fastify.register(fastifyStatic, {
-    // the reply decorator has been added by the first plugin registration
-    decorateReply: false,
-    root: path.join(process.cwd(), "public"),
-    prefix: "/",
-    wildcard: false,
-    maxAge: 3600,
-    dotfiles: "allow",
-  });
-
-  let requestHandler = createRequestHandler({
-    build: fullOptions.build,
-    mode: fullOptions.mode,
-  });
-
-  fastify.all("*", requestHandler);
-};
 
 /**
  * A function that returns the value to use as `context` in route `loader` and
@@ -109,10 +41,10 @@ export function createRequestHandler({
   build: ServerBuild;
   getLoadContext?: GetLoadContextFunction;
   mode?: string;
-}): RequestHandler {
+}) {
   let handleRequest = createRemixRequestHandler(build, mode);
 
-  return async (request, reply) => {
+  return async (request: FastifyRequest, reply: FastifyReply) => {
     let abortController = new AbortController();
     let remixRequest = createRemixRequest(request, abortController);
     let loadContext =
@@ -199,8 +131,3 @@ export function sendRemixResponse(
     reply.send();
   }
 }
-
-export const remixFastifyPlugin = fp(remixFastify, {
-  name: "remix-fastify",
-  fastify: "^3.0.0",
-});
